fix: Return response headers independently on returning response content
The existing option -i (includeGHeaders) enabled this feature.

Headers were returned only with the response content earlier. It can be
convenient not to buffer the whole response, if only headers are needed.

This may be considered both a feature and a breaking change. However,
the headers do not mean adding an excessive content to the result
object.