fix: Adapt http2 connection for Node.js 8.11.2 and Node.js 10
Node.js 8.11.2 and Node.js 10 do not trigger the "close" event on
the socket any more. Also, there is new method "close" on the http2
client, which replaces the earlier (still existing) "destroy".
The complete end of the end of the http2 communication can be caught
only in the callback of the "close" or "destroy" methods and not in
the "close" event handler on the socket any more.