Fix immediate timeouts when specifying a timeout/--connect-timeout argument. The `setTimeout(...)` that has been removed was cancelling requests _immediately_ if a timeout was specified.
It seems as if the setTimeout was added to appease a failing test,
though the test in question claims to "test timed out connection to an
unreachable host". If connecting to an unreachable host the error thrown
is `EHOSTUNREACH`, not `ETIMEDOUT`. Perhaps this was different in versions
of Node < 10, but considering Node v8 is EOL it seems that the removed
test is the bug and not the implementation..
Soo, I increased the default response time of the mock server and
lowered the timeout of the test. It _seems_ to work; YMMV